xds: Merge ClusterResolverLB into CdsLB2
This is a cleanup from gRFC A74:

> The xds_cluster_resolver LB policy will be removed completely, as
> obtaining endpoint addresses will now be done in the xds resolver. The
> code for generating the child policy configs for the priority policy
> will now be done in the cds LB policy.

Since XdsDependencyManager handles the actual resolution,
ClusterResolverLB was only handling config conversion. We just do that
in CdsLB2 now.

The large blob of code from ClusterResolverLB was moved into CdsLB2
unchanged. The cluster resolver tests were left in-place, though, as
they did need changes and it'd be very hard to see the changes if moved
at the same time.
