+++ xds/src/main/java/io/grpc/xds/XdsLbPolicies.java
@@ -19,7 +19,6 @@ package io.grpc.xds;
 final class XdsLbPolicies {
   static final String CLUSTER_MANAGER_POLICY_NAME = "cluster_manager_experimental";
   static final String CDS_POLICY_NAME = "cds_experimental";
-  static final String CLUSTER_RESOLVER_POLICY_NAME = "cluster_resolver_experimental";
   static final String PRIORITY_POLICY_NAME = "priority_experimental";
   static final String CLUSTER_IMPL_POLICY_NAME = "cluster_impl_experimental";
   static final String WEIGHTED_TARGET_POLICY_NAME = "weighted_target_experimental";
