+++ xds/src/main/resources/META-INF/services/io.grpc.LoadBalancerProvider
@@ -2,7 +2,6 @@ io.grpc.xds.CdsLoadBalancerProvider
 io.grpc.xds.PriorityLoadBalancerProvider
 io.grpc.xds.WeightedTargetLoadBalancerProvider
 io.grpc.xds.ClusterManagerLoadBalancerProvider
-io.grpc.xds.ClusterResolverLoadBalancerProvider
 io.grpc.xds.ClusterImplLoadBalancerProvider
 io.grpc.xds.LeastRequestLoadBalancerProvider
 io.grpc.xds.RingHashLoadBalancerProvider
